requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.2
distro==1.8.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.336
langsmith==0.0.64
llvmlite==0.41.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai==1.1.1
openai-whisper==20231106
packaging==23.2
Pillow==10.1.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
soundfile==0.12.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
torch==2.2.0.dev20231107
torchaudio==2.2.0.dev20231107
torchvision==0.17.0.dev20231107
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
